Move searchContacts to ContactController
