+++ app.js
@@ -3,10 +3,6 @@ const express = require('express');
 const logger = require('morgan');
 const cors = require('cors');
 const Sequelize = require('sequelize');
-const parseISO = require('date-fns/parseISO');
-const startOfDay = require('date-fns/startOfDay');
-const endOfDay = require('date-fns/endOfDay');
-const isValid = require('date-fns/isValid');
 
 const app = express();
 
@@ -29,6 +25,8 @@ const sequelize = new Sequelize('hrmdb', 'hrm', pass, {
   host,
   dialect: 'postgres',
 });
+const ContactController = require('./controllers/contact-controller')(sequelize);
+
 console.log('After connect attempt');
 const Contact = require('./models/contact.js')(sequelize, Sequelize);
 
@@ -72,163 +70,6 @@ function redact(form) {
   };
 }
 
-function checkSearchParams(body, res) {
-  const {
-    helpline,
-    firstName,
-    lastName,
-    counselor,
-    phoneNumber,
-    dateFrom,
-    dateTo,
-    singleInput,
-  } = body;
-
-  /**
-   * Prettier currently enforces the operators (eg. ||) to the at the end of the line.
-   * It's much more readable when it's placed at the beggining of the line. But that's
-   * a current limitation of Prettier. There's a PR in place that will fix this in future
-   * versions of Prettier: https://github.com/prettier/prettier/pull/7111.
-   */
-  const anyValue =
-    helpline ||
-    firstName ||
-    lastName ||
-    counselor ||
-    phoneNumber ||
-    dateFrom ||
-    dateTo ||
-    singleInput;
-
-  if (!anyValue) {
-    res.json([]);
-  }
-}
-
-function buildSearchQueryObject(body) {
-  const {
-    helpline,
-    firstName,
-    lastName,
-    counselor,
-    phoneNumber,
-    dateFrom,
-    dateTo,
-    singleInput,
-  } = body;
-
-  const operator = singleInput ? Op.or : Op.and;
-  const isSingleInputValidDate = singleInput && isValid(parseISO(singleInput));
-  const compareDateFrom = dateFrom && !singleInput;
-  const compareDateTo = dateTo && !singleInput;
-
-  return {
-    where: {
-      [Op.and]: [
-        helpline && {
-          helpline,
-        },
-        {
-          [operator]: [
-            (firstName || singleInput) && {
-              'rawJson.childInformation.name.firstName': {
-                [Op.iLike]: singleInput || firstName,
-              },
-            },
-            (lastName || singleInput) && {
-              'rawJson.childInformation.name.lastName': {
-                [Op.iLike]: singleInput || lastName,
-              },
-            },
-            counselor && {
-              twilioWorkerId: counselor,
-            },
-            (phoneNumber || singleInput) && {
-              number: {
-                [Op.iLike]: `%${singleInput || phoneNumber}%`,
-              },
-            },
-            compareDateFrom && {
-              createdAt: {
-                [Op.gte]: startOfDay(parseISO(dateFrom)),
-              },
-            },
-            compareDateTo && {
-              createdAt: {
-                [Op.lte]: endOfDay(parseISO(dateTo)),
-              },
-            },
-            isSingleInputValidDate && {
-              [Op.and]: [
-                {
-                  createdAt: {
-                    [Op.gte]: startOfDay(parseISO(singleInput)),
-                  },
-                },
-                {
-                  createdAt: {
-                    [Op.lte]: endOfDay(parseISO(singleInput)),
-                  },
-                },
-              ],
-            },
-          ],
-        },
-      ],
-    },
-  };
-}
-
-function isNullOrEmptyObject(obj) {
-  return obj == null || Object.keys(obj).length === 0;
-}
-
-function isValidContact(contact) {
-  return (
-    contact &&
-    contact.rawJson &&
-    !isNullOrEmptyObject(contact.rawJson.callType) &&
-    !isNullOrEmptyObject(contact.rawJson.childInformation) &&
-    !isNullOrEmptyObject(contact.rawJson.callerInformation) &&
-    !isNullOrEmptyObject(contact.rawJson.caseInformation)
-  );
-}
-
-function convertContactsToSearchResults(contacts) {
-  return contacts
-    .map(contact => {
-      if (!isValidContact(contact)) {
-        const contactJson = JSON.stringify(contact);
-        console.log(`Invalid Contact: ${contactJson}`);
-        return null;
-      }
-
-      const contactId = contact.id;
-      const dateTime = contact.createdAt;
-      const name = `${contact.rawJson.childInformation.name.firstName} ${contact.rawJson.childInformation.name.lastName}`;
-      const customerNumber = contact.number;
-      const { callType } = contact.rawJson;
-      const categories = 'TBD';
-      const counselor = contact.twilioWorkerId;
-      const notes = contact.rawJson.caseInformation.callSumary;
-
-      return {
-        contactId,
-        overview: {
-          dateTime,
-          name,
-          customerNumber,
-          callType,
-          categories,
-          counselor,
-          notes,
-        },
-        details: redact(contact.rawJson),
-      };
-    })
-    .filter(contact => contact);
-}
-
 // run with node app.js and hit curl localhost:8080/contacts/
 app.get('/contacts', (req, res) => {
   checkAuthentication(req, res);
@@ -260,15 +101,15 @@ app.get('/contacts', (req, res) => {
   });
 });
 
-app.post('/contacts/search', (req, res) => {
+app.post('/contacts/search', async (req, res) => {
   checkAuthentication(req, res);
-  checkSearchParams(req.body, res);
 
-  const queryObject = buildSearchQueryObject(req.body);
-
-  Contact.findAll(queryObject)
-    .then(contacts => res.json(convertContactsToSearchResults(contacts)))
-    .catch(error => console.log(`request rejected: ${error}`));
+  try {
+    const searchResults = await ContactController.searchContacts(req.body);
+    res.json(searchResults);
+  } catch (error) {
+    console.log(`[ContactController.searchContacts]: ${error}`);
+  }
 });
 
 // example: curl -XPOST -H'Content-Type: application/json' localhost:3000/contacts -d'{"hi": 2}'
